Added postInitialize app method Fixed a bunch of warnings

+++ AppKits/AppKit_G3D9/source/MVREngineG3D9.cpp
@@ -50,7 +50,7 @@ void MVREngineG3D9::setupWindowsAndViewports()
 	glm::mat4 initialHeadFrame = _configMap->get("InitialHeadFrame", glm::mat4(1.0));
 	
 	// InterOcularDistance defaults to 2.5 inches (0.2083 ft). This assumes your coordinate system is in feet.
-	float interOcularDistance = _configMap->get("InterOcularDistance", 0.2083);
+	double interOcularDistance = _configMap->get("InterOcularDistance", 0.2083);
 
 	// Each window will have its own OpenGL graphics context and run in its own thread.
 	// Many applications will only have 1 window, but for fastest performance in a CAVE
@@ -120,8 +120,8 @@ void MVREngineG3D9::setupWindowsAndViewports()
 			
 			int width    = _configMap->get(viewportStr + "Width", wSettings->width);
 			int height   = _configMap->get(viewportStr + "Height", wSettings->height);
-			int x        = _configMap->get(viewportStr + "X", 0.0);
-			int y        = _configMap->get(viewportStr + "Y", 0.0);
+			int x        = _configMap->get(viewportStr + "X", 0);
+			int y        = _configMap->get(viewportStr + "Y", 0);
 			wSettings->viewports.push_back(MinVR::Rect2D::xywh(x,y,width,height));
 
 			std::string cameraStr = _configMap->get(viewportStr + "CameraType", "OffAxis");
@@ -130,8 +130,8 @@ void MVREngineG3D9::setupWindowsAndViewports()
 				glm::vec3 topRight = _configMap->get(viewportStr + "TopRight", glm::vec3(1.0, 1.0, 0.0));
 				glm::vec3 botLeft  = _configMap->get(viewportStr + "BotLeft", glm::vec3(-1.0, -1.0, 0.0));
 				glm::vec3 botRight = _configMap->get(viewportStr + "BotRight", glm::vec3(1.0, -1.0, 0.0));
-				float nearClip = _configMap->get(viewportStr + "NearClip", 0.01);
-				float farClip  = _configMap->get(viewportStr + "FarClip", 1000.0);
+				float nearClip = _configMap->get(viewportStr + "NearClip", 0.01f);
+				float farClip  = _configMap->get(viewportStr + "FarClip", 1000.0f);
 				AbstractCameraRef cam(new OffAxisCameraG3D9(topLeft, topRight, botLeft, botRight, initialHeadFrame, interOcularDistance, nearClip, farClip));
 				cameras.push_back(cam);
 			}
